Made it easier to run
Instead of having a nodejs server processing the react/babel then hosting it on a local server, added a file so it can all be done in browser. This is slower but easier for development. Now you simply have to run web.py and go to local port 5000 to see the web app.